add collision improvements
collision checks are handled using configuration spaces (paddles are
inflated by the radius of the ball and the ball is treated as points
instead).

while the axis aligned collisions allow for a pure negation of one of
the axes in the velocity vectors, collisions at the corner (non-axis
aligned) require proper reflection physics. this implements the vector
representation of snell's law.

while it feels like a bit of a hack, i store the previous position of
the ball as that is useful for determining the intersection point
between the ball and the corners of the paddles. this could certainly
be refactored in a way where we precompute the collisions before
moving the ball.

simplifications made:
 - only allow one collision per paddle for the ball. this should not
   matter since the paddle cannot move faster than the ball.
 - no interpoliation for collision time is done. there is a small
   amount of "teleporting" upon collision but in most cases it is
   subpixel.
 - for collisions on the faces whose normals are coincident to the
   velocity vector of the paddle, the velocity vector of the paddle is
   added to that of the ball to counteract the ball flipping "into"
   the paddle. this allows for reflections where the paddle and the
   ball are moving in the same direction.
 - for the collisions with corners, the position is not corrected.

as a general thought, i think treating all players as having a
velocity vector vx and vy is a more natural way of dealing with this
rather than having the `.vertical` and a single `.v` for velocity.